add useTTSState hook
